feat:북마크 삭제 기능 추가

+++ controller/Cmain.js
@@ -473,15 +473,37 @@ exports.profileUpdate = async (req, res) => {
 };
 
 // 북마크 삭제
-exports.bookmarkDelete(async (req, res) => {
+exports.bookmarkDelete = async (req, res) => {
     try {
-        // 선택한 북마크 id 추출
-        // const userId = decodedToken.id;
+        // 응답된 params 경로에서 bookmarkId값 추출
+        const { bookmarkId } = req.params;
+        // 요청 헤더에서 Authorization 값 추출, (bearer[token] 형식)
+        const tokenWithBearer = req.headers.authorization;
+        // bearer와 token을 공백으로 분리해서 실제 토큰만 token 변수에 담음
+        const token = tokenWithBearer.split(" ")[1];
+        // 토큰 없는 경우
+        if (!token) {
+            return res.status(401).send("로그인이 필요합니다.");
+        }
+        // 토큰 디코드하고 디코드된 토큰에서 사용자 id 추출
+        const decodedToken = jwt.verify(token, SECRET);
+        const userId = decodedToken.id;
+        // Bookmarks 모델에서 bookmarks가 위에서 응답받은 bookmarks인지, userid인지 확인
+        const checkBookmark = await models.Bookmarks.findOne({
+            where: { bookmarkId: bookmarkId, id: userId },
+        });
+        // 북마크 id와 사용자 id 일치하지 않는 경우
+        if (!checkBookmark) {
+            return res.status(404).json({ error: "북마크가 없습니다." });
+        }
         const isDeleted = await models.Bookmarks.destroy({
-            where: { id: bookmarkId },
+            where: { bookmarkId: bookmarkId },
         });
         if (isDeleted) {
             res.send({ msg: "북마크가 삭제되었습니다." });
         }
-    } catch (err) {}
-});
+    } catch (err) {
+        console.error("오류 상세 정보:", err);
+        return res.status(500).send("서버 에러, 상세 정보를 확인하세요.");
+    }
+};
